# application/services/custom_services/ticket/src/models/tickets.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const ticketsSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   ticketname: String,
   ticketdescription: String,
   servicetype: { type: Schema.Types.String, ref: 'servicetype' },
   severity: { type: Schema.Types.String, ref: 'severity' }
})

const ticketsModel = mongoose.model('tickets', ticketsSchema, 'tickets');
export default ticketsModel;